+++ src/interactions/types/mod.ts
@@ -2,5 +2,5 @@ export * from "./embed.ts";
 export * from "./interactions.ts";
 export * from "./misc.ts";
 export * from "./slash.ts";
-export * from "./user.ts";
+export * from "./member.ts";
 export * from "./webhook.ts";
